+++ src/main/java/subway/controller/StationController.java
@@ -1,10 +1,19 @@
-package subway;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+package subway.controller;
 
 import java.net.URI;
 import java.util.List;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import subway.service.StationService;
+import subway.service.dto.StationRequest;
+import subway.service.dto.StationResponse;
 
 @RestController
 public class StationController {
