+++ src/test/java/subway/study/RestAssuredTest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.study;
 
 import static org.assertj.core.api.Assertions.assertThat;
